requirements.txt
flask
PyMySQL
authlib
python-dotenv
flask_pymongo
pymongo
bson
flask_cors
dnspython
